how to compare two lists together
This was a challenge where we  have to first convert the .txt file  data into list and then compare two of such list with each other and fo a new list which have only num which are common to both the list.

This comparision can easily be made using list comprehension.